Create publish form component, fix markdown editor, fix prefill
