+++ ui/js/component/publishForm/index.js
@@ -1,0 +1,5 @@
+import React from "react";
+import { connect } from "react-redux";
+import PublishForm from "./view";
+
+export default connect()(PublishForm);
